feat : db connected
db connected and working, solved routing problems
