+++ views/Index.html
@@ -377,16 +377,11 @@
                                 <!-- OM YAHA  DAAL FORM  -->
                                 <div class="tab-pane fade" id="custom-tabs-two-profile" role="tabpanel"
                                     aria-labelledby="custom-tabs-two-profile-tab">
-                                    Mauris tincidunt mi at erat gravida, eget tristique urna bibendum. Mauris pharetra
-                                    purus ut
-                                    ligula tempor, et vulputate metus facilisis. Lorem ipsum dolor sit amet, consectetur
-                                    adipiscing elit. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices
-                                    posuere
-                                    cubilia Curae; Maecenas sollicitudin, nisi a luctus interdum, nisl ligula placerat
-                                    mi, quis
-                                    posuere purus ligula eu lectus. Donec nunc tellus, elementum sit amet ultricies at,
-                                    posuere
-                                    nec nunc. Nunc euismod pellentesque diam.  CHOMU 
+                                    <form method="post" action="/register">
+                                        <input type="text" name="fname" placeholder="first name">
+                                        <input type="text" name="lname" placeholder="last name">
+                                        <button type="submit">send</button>
+                                    </form>
                                 </div>
                                 <div class="tab-pane fade" id="custom-tabs-two-messages" role="tabpanel"
                                     aria-labelledby="custom-tabs-two-messages-tab">
